# Strings are not the fastpath