@mattfarina 2016-04-05
@sdboyer great catch. Benchmarks for the win!